1.1.1 - Alterado o tipo primitivo do cpf e celular para que fosse possível armazena-los em banco e em variavel no Software

+++ BarbeariaBanco.sql
@@ -5,10 +5,10 @@ CREATE TABLE `pessoa` (
   `ID_PESSOA` int NOT NULL AUTO_INCREMENT,
   `NOME` varchar(45) NOT NULL,
   `SOBRENOME` varchar(45) NOT NULL,
-  `CPF` int NOT NULL,
+  `CPF` long NOT NULL,
   `SEXO` varchar(45) NOT NULL,
   `EMAIL` varchar(45) NOT NULL,
-  `CELULAR` int NOT NULL,
+  `CELULAR` long NOT NULL,
   `NASCIMENTO` date NOT NULL,
   `SENHA` varchar(100) NOT NULL,
   PRIMARY KEY (`ID_PESSOA`));
